+++ src/main/java/mekanism/client/gui/GuiFormulaicAssemblicator.java
@@ -152,7 +152,7 @@ public class GuiFormulaicAssemblicator extends GuiMekanismTile<TileEntityFormula
         }
 
         if (tileEntity.operatingTicks > 0) {
-            int display = (int) ((double) tileEntity.operatingTicks * 22 / (double) tileEntity.ticksRequired);
+            int display = (int) ((double) tileEntity.operatingTicks * 22 / tileEntity.ticksRequired);
             drawTexturedModalRect(guiWidth + 86, guiHeight + 43, 176, 48, display, 16);
         }
