+++ src/main/java/mekanism/client/render/tileentity/RenderEnergyCube.java
@@ -97,7 +97,7 @@ public class RenderEnergyCube extends TileEntitySpecialRenderer<TileEntityEnergy
 
             GlStateManager.pushMatrix();
             GlStateManager.scale(0.4F, 0.4F, 0.4F);
-            GL11.glColor4f((float) c[0] / 255F, (float) c[1] / 255F, (float) c[2] / 255F,
+            GL11.glColor4f(c[0] / 255F, c[1] / 255F, c[2] / 255F,
                   (float) (tileEntity.getEnergy() / tileEntity.getMaxEnergy()));
             GlStateManager
                   .translate(0, (float) Math.sin(Math.toRadians((MekanismClient.ticksPassed + partialTick) * 3)) / 7,
