+++ src/main/java/mekanism/common/block/BlockGasTank.java
@@ -98,7 +98,7 @@ public class BlockGasTank extends BlockContainer {
           ItemStack stack) {
         TileEntityBasicBlock tileEntity = (TileEntityBasicBlock) world.getTileEntity(pos);
 
-        int side = MathHelper.floor((double) (placer.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+        int side = MathHelper.floor(placer.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
         int change = 3;
 
         switch (side) {
