+++ src/main/java/mekanism/common/item/ItemBlockGasTank.java
@@ -305,7 +305,7 @@ public class ItemBlockGasTank extends ItemBlock implements IGasItem, ISustainedI
 
     @Override
     public double getDurabilityForDisplay(ItemStack stack) {
-        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / (double) getMaxGas(stack));
+        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / getMaxGas(stack));
     }
 
     @Override
