+++ src/main/java/mekanism/common/item/ItemJetpack.java
@@ -52,7 +52,7 @@ public class ItemJetpack extends ItemArmor implements IGasItem, ISpecialArmor {
 
     @Override
     public double getDurabilityForDisplay(ItemStack stack) {
-        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / (double) getMaxGas(stack));
+        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / getMaxGas(stack));
     }
 
     @Override
