+++ src/main/java/mekanism/common/item/ItemScubaTank.java
@@ -63,7 +63,7 @@ public class ItemScubaTank extends ItemArmor implements IGasItem {
 
     @Override
     public double getDurabilityForDisplay(ItemStack stack) {
-        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / (double) getMaxGas(stack));
+        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / getMaxGas(stack));
     }
 
     @Override
