+++ src/main/java/mekanism/generators/common/item/ItemHohlraum.java
@@ -102,7 +102,7 @@ public class ItemHohlraum extends ItemMekanism implements IGasItem {
 
     @Override
     public double getDurabilityForDisplay(ItemStack stack) {
-        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / (double) getMaxGas(stack));
+        return 1D - ((getGas(stack) != null ? (double) getGas(stack).amount : 0D) / getMaxGas(stack));
     }
 
     @Override
